Fix some data types that were not applied correctly

+++ strengthjournal-server/StrengthJournal.DataAccess/Model/WorkoutLogEntry.cs
@@ -15,7 +15,7 @@ namespace StrengthJournal.DataAccess.Model
         public DateTime EntryDateUTC { get; set; }
         public User User { get; set; }
         public ICollection<WorkoutLogEntrySet> Sets { get; set; } = new List<WorkoutLogEntrySet>();
-        public uint? BodyWeightPIT { get; set; }
+        public decimal? BodyWeightPIT { get; set; }
         public WeightUnit? BodyWeightPITUnit { get; set; }
         public string Notes { get; set; } = string.Empty;
     }
